% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UN.SDG.03-package.R
\docType{package}
\name{UN.SDG.03-package}
\alias{UN.SDG.03}
\alias{UN.SDG.03-package}
\title{UN.SDG.03: UN Sustainable Development Goal 3: Health and Well-Being (Update: 12 September 2023)}
\description{
UN Sustainable Development Goal 3 (Health and Well-being) Indicators Database archived on 12 September 2023.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramiromagno/UN.SDG.03}
  \item \url{https://rmagno.eu/UN.SDG.03/}
  \item Report bugs at \url{https://github.com/ramiromagno/UN.SDG.03/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
